error while using say fixed
Fixed an error where if no message parsed after say bot would try and send it anyways.

+++ commands/fun/say.cpscmd/s.js
@@ -1,6 +1,6 @@
 module.exports = {
   name: 's',
-  func(msg, { send, author, suffix, Constants }) {
+  func(msg, { send, reply, author, suffix, Constants }) {
     if (author.id === Constants.users.WILLYZ ||
       author.id === Constants.users.EVILDEATHPRO ||
       author.id === Constants.users.LUCAS ||
@@ -9,6 +9,7 @@ module.exports = {
     ) {
       // Antonio, 166..12
 
+      if(suffix === "") return reply("Please include a message for the bot to repeat");
       msg.delete().catch(_ => _);
       return send(suffix, { disableEveryone: true });
     }
